refactor(lotto): Rank를 model로 패키지 이동
변경 전
- constants 패키지

변경 후
- model 패키지
