+++ src/main/java/lotto/controller/LottoSalesController.java
@@ -1,6 +1,6 @@
 package lotto.controller;
 
-import lotto.constants.Rank;
+import lotto.model.Rank;
 import lotto.model.BonusNumber;
 import lotto.model.Lotto;
 import lotto.model.Amount;
